mt6895-common: udfps: Defer extCmd until fod ui is ready
This does not work with xiaomi's original display driver,
since partial reads from it are broken and it rejects all
reads whose size is insufficient to get the whole response.
However, the size is not known in prior, so we actually
want to read the header first to determine the actual size
and continue reading then. This requires a kernel side fix
like [1].

To make this work with the stock kernel driver instead,
adjust the parseDispEvent method to read with a larger
buffer directly and decrease the size of the data
according to the header later.

[1]: https://github.com/cupid-development/android_kernel_xiaomi_sm8450-modules/commit/b80cb5a0289b3f37217939123087d7d45354bed4

Change-Id: Id8f9ac8ef32462dd00832af12ff60a3610103e2a